backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
alembic
python-multipart
python-jose[cryptography]
passlib[argon2]
